convert new-blocks-test to typescript
